requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.0
ollama==0.1.4
deepseek==0.0.1
